[Impeller] use explicit half types

+++ impeller/entity/shaders/atlas_fill.vert
@@ -2,17 +2,19 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
+#include <impeller/types.glsl>
+
 uniform VertInfo {
   mat4 mvp;
 }
 vert_info;
 
-in vec2 position;
-in vec2 texture_coords;
-in vec4 color;
+in f16vec2 position;
+in f16vec2 texture_coords;
+in f16vec4 color;
 
-out vec2 v_texture_coords;
-out vec4 v_color;
+out f16vec2 v_texture_coords;
+out f16vec4 v_color;
 
 void main() {
   gl_Position = vert_info.mvp * vec4(position, 0.0, 1.0);
