+++ impeller/entity/shaders/blending/blend.frag
@@ -2,18 +2,19 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
+#include <impeller/types.glsl>
 #include <impeller/texture.glsl>
 
 uniform sampler2D texture_sampler_src;
 
 uniform FragInfo {
-  float texture_sampler_y_coord_scale;
-  float input_alpha;
+  float16_t texture_sampler_y_coord_scale;
+  float16_t input_alpha;
 } frag_info;
 
-in vec2 v_texture_coords;
+in f16vec2 v_texture_coords;
 
-out vec4 frag_color;
+out f16vec4 frag_color;
 
 void main() {
   frag_color = IPSample(texture_sampler_src, v_texture_coords,
