+++ impeller/entity/shaders/gaussian_blur.vert
@@ -2,17 +2,19 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
+#include <impeller/types.glsl>
+
 uniform FrameInfo {
   mat4 mvp;
 }
 frame_info;
 
 in vec2 vertices;
-in vec2 texture_coords;
-in vec2 src_texture_coords;
+in f16vec2 texture_coords;
+in f16vec2 src_texture_coords;
 
-out vec2 v_texture_coords;
-out vec2 v_src_texture_coords;
+out f16vec2 v_texture_coords;
+out f16vec2 v_src_texture_coords;
 
 void main() {
   gl_Position = frame_info.mvp * vec4(vertices, 0.0, 1.0);
