+++ impeller/entity/shaders/position.vert
@@ -2,16 +2,17 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
+#include <impeller/types.glsl>
 #include <impeller/transform.glsl>
 
 uniform VertInfo {
   mat4 mvp;
-  vec4 color;
+  f16vec4 color;
 } vert_info;
 
-in vec2 position;
+in f16vec2 position;
 
-out vec4 v_color;
+out f16vec4 v_color;
 
 void main() {
   gl_Position = vert_info.mvp * vec4(position, 0.0, 1.0);
