+++ impeller/entity/shaders/position_color.vert
@@ -2,6 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
+#include <impeller/types.glsl>
 #include <impeller/transform.glsl>
 
 uniform VertInfo {
@@ -9,9 +10,9 @@ uniform VertInfo {
 } vert_info;
 
 in vec2 position;
-in vec4 color;
+in f16vec4 color;
 
-out vec4 v_color;
+out f16vec4 v_color;
 
 void main() {
   gl_Position = vert_info.mvp * vec4(position, 0.0, 1.0);
